// cluster/accessportal/accessportal/web/package/src/apis/metav1/metav1.ts
export interface ObjectReference {
    /**
     * APIVersion is the API version of the reference resource
     *
     * @generated from protobuf field: string apiVersion = 1
     */
    apiVersion: string;
    /**
     * Kind is the resource kind. A kind is unique per an APIVersion.
     *
     * @generated from protobuf field: string kind = 2
     */
    kind: string;
    /**
     * UID is the resource object UID. A UID is unique across the entire Cluster.
     *
     * @generated from protobuf field: string uid = 3
     */
    uid: string;
    /**
     * Name is the resource object name. A name is unique per a certain kind.
     *
     * @generated from protobuf field: string name = 4
     */
    name: string;
    /**
     * ResourceVersion is the resourceVersion of the resource.
     *
     * @generated from protobuf field: string resourceVersion = 5
     */
    resourceVersion: string;
}
/**
 * Duration is a length of time that is expressed in a single unit of choice.
 * Exactly one of the units is set (e.g. `seconds: 30`).
 *
 * @generated from protobuf message octelium.api.main.meta.v1.Duration
 */
export interface Duration {
    /**
     * Type sets the unit in which the Duration is expressed
     *
     * @generated from protobuf oneof: type
     */
    type: {
        oneofKind: "milliseconds";
        /**
         * Milliseconds is the duration set in milliseconds
         *
         * @generated from protobuf field: uint32 milliseconds = 1
         */
        milliseconds: number;
    } | {
        oneofKind: "seconds";
        /**
         * Seconds is the duration set in seconds
         *
         * @generated from protobuf field: uint32 seconds = 2
         */
        seconds: number;
    } | {
        oneofKind: "minutes";
        /**
         * Minutes is the duration set in minutes
         *
         * @generated from protobuf field: uint32 minutes = 3
         */
        minutes: number;
    } | {
        oneofKind: "hours";
        /**
         * Hours is the duration set in hours
         *
         * @generated from protobuf field: uint32 hours = 4
         */
        hours: number;
    } | {
        oneofKind: "days";
        /**
         * Days is the duration set in days
         *
         * @generated from protobuf field: uint32 days = 5
         */
        days: number;
    } | {
        oneofKind: "weeks";
        /**
         * Weeks is the duration set in weeks
         *
         * @generated from protobuf field: uint32 weeks = 6
         */
        weeks: number;
    } | {
        oneofKind: "months";
        /**
         * Months is the duration set in months
         *
         * @generated from protobuf field: uint32 months = 7
         */
        months: number;
    } | {
        oneofKind: undefined;
    };
}
// ...
class ObjectReference$Type extends MessageType<ObjectReference> {
    constructor() {
        super("octelium.api.main.meta.v1.ObjectReference", [
            { no: 1, name: "apiVersion", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "kind", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "uid", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "resourceVersion", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<ObjectReference>): ObjectReference {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.apiVersion = "";
        message.kind = "";
        message.uid = "";
        message.name = "";
        message.resourceVersion = "";
        if (value !== undefined)
            reflectionMergePartial<ObjectReference>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ObjectReference): ObjectReference {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string apiVersion */ 1:
                    message.apiVersion = reader.string();
                    break;
                case /* string kind */ 2:
                    message.kind = reader.string();
                    break;
                case /* string uid */ 3:
                    message.uid = reader.string();
                    break;
                case /* string name */ 4:
                    message.name = reader.string();
                    break;
                case /* string resourceVersion */ 5:
                    message.resourceVersion = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ObjectReference, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string apiVersion = 1; */
        if (message.apiVersion !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.apiVersion);
        /* string kind = 2; */
        if (message.kind !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.kind);
        /* string uid = 3; */
        if (message.uid !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.uid);
        /* string name = 4; */
        if (message.name !== "")
            writer.tag(4, WireType.LengthDelimited).string(message.name);
        /* string resourceVersion = 5; */
        if (message.resourceVersion !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.resourceVersion);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.meta.v1.ObjectReference
 */
export const ObjectReference = new ObjectReference$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Duration$Type extends MessageType<Duration> {
    constructor() {
        super("octelium.api.main.meta.v1.Duration", [
            { no: 1, name: "milliseconds", kind: "scalar", oneof: "type", T: 13 /*ScalarType.UINT32*/ },
            { no: 2, name: "seconds", kind: "scalar", oneof: "type", T: 13 /*ScalarType.UINT32*/ },
            { no: 3, name: "minutes", kind: "scalar", oneof: "type", T: 13 /*ScalarType.UINT32*/ },
            { no: 4, name: "hours", kind: "scalar", oneof: "type", T: 13 /*ScalarType.UINT32*/ },
            { no: 5, name: "days", kind: "scalar", oneof: "type", T: 13 /*ScalarType.UINT32*/ },
            { no: 6, name: "weeks", kind: "scalar", oneof: "type", T: 13 /*ScalarType.UINT32*/ },
            { no: 7, name: "months", kind: "scalar", oneof: "type", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<Duration>): Duration {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.type = { oneofKind: undefined };
        if (value !== undefined)
            reflectionMergePartial<Duration>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Duration): Duration {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* uint32 milliseconds */ 1:
                    message.type = {
                        oneofKind: "milliseconds",
                        milliseconds: reader.uint32()
                    };
                    break;
                case /* uint32 seconds */ 2:
                    message.type = {
                        oneofKind: "seconds",
                        seconds: reader.uint32()
                    };
                    break;
                case /* uint32 minutes */ 3:
                    message.type = {
                        oneofKind: "minutes",
                        minutes: reader.uint32()
                    };
                    break;
                case /* uint32 hours */ 4:
                    message.type = {
                        oneofKind: "hours",
                        hours: reader.uint32()
                    };
                    break;
                case /* uint32 days */ 5:
                    message.type = {
                        oneofKind: "days",
                        days: reader.uint32()
                    };
                    break;
                case /* uint32 weeks */ 6:
                    message.type = {
                        oneofKind: "weeks",
                        weeks: reader.uint32()
                    };
                    break;
                case /* uint32 months */ 7:
                    message.type = {
                        oneofKind: "months",
                        months: reader.uint32()
                    };
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Duration, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* uint32 milliseconds = 1; */
        if (message.type.oneofKind === "milliseconds")
            writer.tag(1, WireType.Varint).uint32(message.type.milliseconds);
        /* uint32 seconds = 2; */
        if (message.type.oneofKind === "seconds")
            writer.tag(2, WireType.Varint).uint32(message.type.seconds);
        /* uint32 minutes = 3; */
        if (message.type.oneofKind === "minutes")
            writer.tag(3, WireType.Varint).uint32(message.type.minutes);
        /* uint32 hours = 4; */
        if (message.type.oneofKind === "hours")
            writer.tag(4, WireType.Varint).uint32(message.type.hours);
        /* uint32 days = 5; */
        if (message.type.oneofKind === "days")
            writer.tag(5, WireType.Varint).uint32(message.type.days);
        /* uint32 weeks = 6; */
        if (message.type.oneofKind === "weeks")
            writer.tag(6, WireType.Varint).uint32(message.type.weeks);
        /* uint32 months = 7; */
        if (message.type.oneofKind === "months")
            writer.tag(7, WireType.Varint).uint32(message.type.months);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.meta.v1.Duration
 */
export const Duration = new Duration$Type();